Add factory aliasing, with a configuration option.
