+++ CHANGELOG.markdown
@@ -1,7 +1,9 @@
 # Changelog
 
-## 0.2.0 — Unreleased.
+## 0.2.0 — Unreleased
 
+* Add aliasing for factory names.
+* Add configuration for custom alias name.
 * Support multiple versions of factory_bot.
 
 ## 0.1.0 — November 5th, 2021
